GC: Avoid errors when deleting non-empty folders
I88304d34c and Ia555bce00 modified the way errors are handled when
trying to delete non-empty reference folders. Before, this error was
silently ignored as it was considered an expected output. Now, every
failed folder delete is logged which can be noisy.

Instead, verify the folder is empty just before deleting it. This way,
the obvious error when it is not is ignored but any other error avoiding
deletion of an eligible folder is logged.

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/file/GC.java
@@ -908,10 +908,22 @@ public class GC {
 
 	private void delete(Path d) {
 		try {
-			Files.delete(d);
+			if (isEmpty(d)) {
+				Files.delete(d);
+			}
 		} catch (IOException e) {
 			LOG.error(MessageFormat.format(JGitText.get().cannotDeleteFile, d),
 					e);
+		}
+	}
+
+	private boolean isEmpty(Path p) {
+		try (DirectoryStream<Path> dirStream = Files.newDirectoryStream(p)) {
+			boolean hasNext = dirStream.iterator().hasNext();
+			return !hasNext;
+		} catch (IOException e) {
+			// If not able to check, return false
+			return false;
 		}
 	}
